HW1/requirements.txt
# Core ML libraries
torch
torchaudio
torchvision

# Audio processing
librosa
soundfile

# Traditional ML
scikit-learn
numpy
pandas

# Visualization
matplotlib
seaborn

# Experiment tracking
wandb

# Configuration management
hydra-core
omegaconf

# Progress bars and utilities
tqdm

# Jupyter
jupyter
ipywidgets